+++ src/theme/BlogFeaturedPosts/index.tsx
@@ -2,6 +2,7 @@ import React from "react"
 import Link from "@docusaurus/Link"
 import type {Props} from "@theme/BlogListPage"
 import {FrontMatter} from "@theme/BlogPostPage"
+import {BlogAuthor} from "../BlogAuthor"
 
 function BlogFeaturedPosts({items}: {items: Props["items"]}): JSX.Element {
   return (
@@ -16,14 +17,7 @@ function BlogFeaturedPosts({items}: {items: Props["items"]}): JSX.Element {
             className="flex flex-col gap-4 text-black !no-underline hover:text-black"
           >
             <div className="flex items-center gap-2">
-              {post.content.metadata.authors[0] && (
-                <img
-                  src={post.content.metadata.authors[0].imageURL}
-                  alt={post.content.metadata.authors[0].name}
-                  className="h-6 w-6 rounded-full"
-                />
-              )}
-              <span className="text-sm font-medium">{post.content.metadata.authors[0]?.name}</span>
+              {post.content.metadata.authors[0] && <BlogAuthor author={post.content.metadata.authors[0]} />}
             </div>
             <div>
               <div className="text-lg font-semibold">{post.content.metadata.title}</div>
